+++ widget/dtype/int32.go
@@ -35,5 +35,5 @@ func (n NullUint32) Value() (driver.Value, error) {
 	if !n.Valid {
 		return nil, nil
 	}
-	return n.Uint32, nil
+	return int64(n.Uint32), nil
 }
